+++ src-opam/dockerfile_distro.ml
@@ -22,9 +22,9 @@ type t = [
   | `Alpine of [ `V3_3 | `V3_4 | `V3_5 | `V3_6 | `V3_7 | `V3_8 | `Latest ]
   | `CentOS of [ `V6 | `V7 | `Latest ]
   | `Debian of [ `V9 | `V8 | `V7 | `Stable | `Testing | `Unstable ]
-  | `Fedora of [ `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `Latest ]
+  | `Fedora of [ `V21 | `V22 | `V23 | `V24 | `V25 | `V26 | `V27 | `V28 | `V29 | `Latest ]
   | `OracleLinux of [ `V7 | `Latest ]
-  | `OpenSUSE of [ `V42_1 | `V42_2 | `V42_3 | `Latest ]
+  | `OpenSUSE of [ `V42_1 | `V42_2 | `V42_3 | `V15_0 | `Latest ]
   | `Ubuntu of [ `V12_04 | `V14_04 | `V15_04 | `V15_10 | `V16_04 | `V16_10 | `V17_04 | `V17_10 | `V18_04 | `V18_10 | `LTS | `Latest ]
 ] [@@deriving sexp]
 
@@ -39,18 +39,18 @@ let distros = [
   `CentOS `V6; `CentOS `V7; `CentOS `Latest;
   `Debian `V9; `Debian `V8; `Debian `V7;
   `Debian `Stable; `Debian `Testing; `Debian `Unstable;
-  `Fedora `V23; `Fedora `V24; `Fedora `V25; `Fedora `V26; `Fedora `V27; `Fedora `V28; `Fedora `Latest;
+  `Fedora `V23; `Fedora `V24; `Fedora `V25; `Fedora `V26; `Fedora `V27; `Fedora `V28; `Fedora `V29; `Fedora `Latest;
   `OracleLinux `V7; `OracleLinux `Latest;
-  `OpenSUSE `V42_1; `OpenSUSE `V42_2; `OpenSUSE `V42_3; `OpenSUSE `Latest;
+  `OpenSUSE `V42_1; `OpenSUSE `V42_2; `OpenSUSE `V42_3; `OpenSUSE `V15_0; `OpenSUSE `Latest;
   `Ubuntu `V12_04; `Ubuntu `V14_04; `Ubuntu `V15_04; `Ubuntu `V15_10;
   `Ubuntu `V16_04; `Ubuntu `V16_10; `Ubuntu `V17_04; `Ubuntu `V17_10; `Ubuntu `V18_04; `Ubuntu `V18_10;
   `Ubuntu `Latest; `Ubuntu `LTS ]
   
 let distro_status (d:t) : status = match d with
   | `Alpine ( `V3_3 | `V3_4 | `V3_5 | `V3_6) -> `Deprecated
-  | `Alpine `V3_7 -> `Active `Tier1
+  | `Alpine `V3_7 -> `Active `Tier2
   | `Alpine `V3_8 -> `Active `Tier1
-  | `Alpine `Latest -> `Alias (`Alpine `V3_7)
+  | `Alpine `Latest -> `Alias (`Alpine `V3_8)
   | `CentOS `V7 -> `Active `Tier2
   | `CentOS `V6 -> `Deprecated
   | `CentOS `Latest -> `Alias (`CentOS `V7)
@@ -61,17 +61,15 @@ let distro_status (d:t) : status = match d with
   | `Debian `Testing -> `Active `Tier2
   | `Debian `Unstable -> `Active `Tier2
   | `Fedora ( `V21 | `V22 | `V23 | `V24 | `V25 | `V26) -> `Deprecated
-  | `Fedora `V28 -> `Active `Tier1
-  | `Fedora `V27 -> `Active `Tier2
-  | `Fedora `Latest -> `Alias (`Fedora `V28)
+  | `Fedora (`V27|`V28|`V29) -> `Active `Tier2
+  | `Fedora `Latest -> `Alias (`Fedora `V29)
   | `OracleLinux `V7 -> `Active `Tier2
   | `OracleLinux `Latest -> `Alias (`OracleLinux `V7)
   | `OpenSUSE `V42_1 | `OpenSUSE `V42_2 -> `Deprecated
-  | `OpenSUSE `V42_3 -> `Active `Tier2
-  | `OpenSUSE `Latest -> `Alias (`OpenSUSE `V42_3)
-  | `Ubuntu `V14_04 -> `Active `Tier2
-  | `Ubuntu ( `V16_04 |  `V18_04) -> `Active `Tier1
-  | `Ubuntu ( `V18_10) -> `Active `Tier2
+  | `OpenSUSE (`V42_3|`V15_0) -> `Active `Tier2
+  | `OpenSUSE `Latest -> `Alias (`OpenSUSE `V15_0)
+  | `Ubuntu (`V14_04 |`V16_04 | `V18_10 ) -> `Active `Tier2
+  | `Ubuntu `V18_04 -> `Active `Tier1
   | `Ubuntu ( `V12_04 | `V15_04 | `V15_10 | `V16_10 | `V17_04 | `V17_10) -> `Deprecated
   | `Ubuntu `LTS -> `Alias (`Ubuntu `V18_04)
   | `Ubuntu `Latest -> `Alias (`Ubuntu `V18_10)
@@ -143,11 +141,13 @@ let builtin_ocaml_of_distro (d:t) : string option =
   |`Fedora `V26 -> Some "4.04.0"
   |`Fedora `V27 -> Some "4.05.0"
   |`Fedora `V28 -> Some "4.06.0"
+  |`Fedora `V29 -> Some "4.07.0"
   |`CentOS `V6 -> Some "3.11.2"
   |`CentOS `V7 -> Some "4.01.0"
   |`OpenSUSE `V42_1 -> Some "4.02.3"
   |`OpenSUSE `V42_2 -> Some "4.03.0"
   |`OpenSUSE `V42_3 -> Some "4.03.0"
+  |`OpenSUSE `V15_0 -> Some "4.05.0"
   |`OracleLinux `V7 -> None
   |`Alpine `Latest |`CentOS `Latest |`OracleLinux `Latest
   |`OpenSUSE `Latest |`Ubuntu `LTS | `Ubuntu `Latest
@@ -185,6 +185,7 @@ let tag_of_distro (d:t) = match d with
   |`Fedora `V26 -> "fedora-26"
   |`Fedora `V27 -> "fedora-27"
   |`Fedora `V28 -> "fedora-28"
+  |`Fedora `V29 -> "fedora-29"
   |`OracleLinux `V7 -> "oraclelinux-7"
   |`OracleLinux `Latest -> "oraclelinux"
   |`Alpine `V3_3 -> "alpine-3.3"
@@ -197,6 +198,7 @@ let tag_of_distro (d:t) = match d with
   |`OpenSUSE `V42_1 -> "opensuse-42.1"
   |`OpenSUSE `V42_2 -> "opensuse-42.2"
   |`OpenSUSE `V42_3 -> "opensuse-42.3"
+  |`OpenSUSE `V15_0 -> "opensuse-15.0"
   |`OpenSUSE `Latest -> "opensuse"
 
 let distro_of_tag x : t option = match x with
@@ -228,6 +230,7 @@ let distro_of_tag x : t option = match x with
   |"fedora-26" -> Some (`Fedora `V26)
   |"fedora-27" -> Some (`Fedora `V27)
   |"fedora-28" -> Some (`Fedora `V28)
+  |"fedora-29" -> Some (`Fedora `V29)
   |"fedora" -> Some (`Fedora `Latest)
   |"oraclelinux-7" -> Some (`OracleLinux `V7)
   |"oraclelinux" -> Some (`OracleLinux `Latest)
@@ -241,6 +244,7 @@ let distro_of_tag x : t option = match x with
   |"opensuse-42.1" -> Some (`OpenSUSE `V42_1)
   |"opensuse-42.2" -> Some (`OpenSUSE `V42_2)
   |"opensuse-42.3" -> Some (`OpenSUSE `V42_3)
+  |"opensuse-15.0" -> Some (`OpenSUSE `V15_0)
   |"opensuse" -> Some (`OpenSUSE `Latest)
   |_ -> None
 
@@ -273,6 +277,7 @@ let rec human_readable_string_of_distro (d:t) =
   |`Fedora `V26 -> "Fedora 26"
   |`Fedora `V27 -> "Fedora 27"
   |`Fedora `V28 -> "Fedora 28"
+  |`Fedora `V29 -> "Fedora 29"
   |`OracleLinux `V7 -> "OracleLinux 7"
   |`Alpine `V3_3 -> "Alpine 3.3"
   |`Alpine `V3_4 -> "Alpine 3.4"
@@ -283,6 +288,7 @@ let rec human_readable_string_of_distro (d:t) =
   |`OpenSUSE `V42_1 -> "OpenSUSE 42.1"
   |`OpenSUSE `V42_2 -> "OpenSUSE 42.2"
   |`OpenSUSE `V42_3 -> "OpenSUSE 42.3"
+  |`OpenSUSE `V15_0 -> "OpenSUSE 15.0 (Leap)"
   |`Alpine `Latest | `Ubuntu `Latest | `Ubuntu `LTS | `CentOS `Latest | `Fedora `Latest
   |`OracleLinux `Latest | `OpenSUSE `Latest -> alias ()